refactor(web): app.js 2,408줄을 12개 ES module로 분해 (Phase 23)
번들러 없이 브라우저 네이티브 ES Modules 만으로 분리. 동작 변경 0
(audio.ended의 imported binding TypeError 수정 1건만 — FM-1 잔존 버그).

도메인 단위 12 모듈:
- state.js (59) — 전역 상태 + setter
- dom.js (74) — getElementById 단일 $ 객체
- util.js (87) — iconFor/formatSize/formatDuration/esc/splitExtension/
  parentDir/rewritePathAfterFolderRename/consumeSSE
- router.js (66) — URL ↔ view + popstate + toolbar wiring
- settings.js (122) — 다운로드 설정 모달
- convert.js (211) — TS→MP4 SSE consumer
- tree.js (211) — 사이드바 폴더 트리 + 모바일 토글
- fileOps.js (369) — upload + folder create + delete + DnD + rename
- browse.js (494) — 디렉토리 조회 + 그리드/테이블 + lightbox + 오디오
- urlImportJobs.js (184) — 잡 부트스트랩 + 구독 + cancel/dismiss
- urlImport.js (693) — 모달 + POST + SSE handler + 배치 상태
- main.js (46) — entry, import + init wiring

순환 의존 회피:
- browse → tree(highlightTreeCurrent), fileOps(attach*Handlers,
  openRenameModal, deleteFile/Folder), convert(openConvertModal),
  router(syncURL) 모두 단방향
- 역방향(tree/fileOps/convert/router → browse, jobs → urlImport↔jobs)은
  wireXxx(deps)·setURLImportDeps()로 late-bound DI

기타:
- 모듈 모드 자동 strict — 'use strict' 제거
- index.html: <script type="module" src="/main.js?v=29">
- 모든 모듈 상단 한 줄 요약 주석
- 각 모듈 ≤ 500 LoC (urlImport.js 693만 700 직전, spec 허용)
- spec/plan/todo는 tasks/ 하위에 동기

검증: go test/vet 통과, JS syntax 12/12 그린, 정적 파일 13/13 200,
spec §5.2 회귀 체크리스트 전체 통과 (FM-1 ~ FM-8 단계마다).

부모 spec: tasks/spec-frontend-modularization.md
플랜: tasks/plan-frontend-modularization.md
