+++ web/dom.js
@@ -1,0 +1,74 @@
+// dom.js — getElementById 결과를 단일 $ 객체로 묶어 export
+
+export const $ = {
+  breadcrumb:           document.getElementById('breadcrumb'),
+  fileList:             document.getElementById('file-list'),
+  uploadZone:           document.getElementById('upload-zone'),
+  fileInput:            document.getElementById('file-input'),
+  uploadProgress:       document.getElementById('upload-progress'),
+  lightbox:             document.getElementById('lightbox'),
+  lbContent:            document.getElementById('lb-content'),
+  lbClose:              document.getElementById('lb-close'),
+  lbPrev:               document.getElementById('lb-prev'),
+  lbNext:               document.getElementById('lb-next'),
+  audioPlayer:          document.getElementById('audio-player'),
+  audioEl:              document.getElementById('audio-el'),
+  audioTitle:           document.getElementById('audio-title'),
+  playlistEl:           document.getElementById('playlist'),
+  newFolderBtn:         document.getElementById('new-folder-btn'),
+  browseSummary:        document.getElementById('browse-summary'),
+  browseToolbar:        document.getElementById('browse-toolbar'),
+  toolbarSearch:        document.getElementById('toolbar-search'),
+  toolbarSort:          document.getElementById('toolbar-sort'),
+  selectAllFiles:       document.getElementById('select-all-files'),
+  selectionSummary:     document.getElementById('selection-summary'),
+  clearSelectionBtn:    document.getElementById('clear-selection-btn'),
+  folderModal:          document.getElementById('folder-modal'),
+  folderNameInput:      document.getElementById('folder-name-input'),
+  folderCancelBtn:      document.getElementById('folder-cancel-btn'),
+  folderConfirmBtn:     document.getElementById('folder-confirm-btn'),
+  folderError:          document.getElementById('folder-error'),
+  sidebar:              document.getElementById('sidebar'),
+  sidebarToggle:        document.getElementById('sidebar-toggle'),
+  sidebarBackdrop:      document.getElementById('sidebar-backdrop'),
+  treeRoot:             document.getElementById('tree-root'),
+  renameModal:          document.getElementById('rename-modal'),
+  renameTitle:          document.getElementById('rename-title'),
+  renameHint:           document.getElementById('rename-hint'),
+  renameInput:          document.getElementById('rename-input'),
+  renameError:          document.getElementById('rename-error'),
+  renameCancelBtn:      document.getElementById('rename-cancel-btn'),
+  renameConfirmBtn:     document.getElementById('rename-confirm-btn'),
+  urlImportBtn:         document.getElementById('url-import-btn'),
+  urlModal:             document.getElementById('url-modal'),
+  urlInput:             document.getElementById('url-input'),
+  urlError:             document.getElementById('url-error'),
+  urlRows:              document.getElementById('url-rows'),
+  urlSummary:           document.getElementById('url-summary'),
+  urlResult:            document.getElementById('url-result'),
+  urlCancelBtn:         document.getElementById('url-cancel-btn'),
+  urlConfirmBtn:        document.getElementById('url-confirm-btn'),
+  urlClearFinishedBtn:  document.getElementById('url-clear-finished-btn'),
+  urlBadge:             document.getElementById('url-badge'),
+  convertAllBtn:        document.getElementById('convert-all-btn'),
+  convertModal:         document.getElementById('convert-modal'),
+  convertFileList:      document.getElementById('convert-file-list'),
+  convertDeleteOrig:    document.getElementById('convert-delete-original'),
+  convertError:         document.getElementById('convert-error'),
+  convertResult:        document.getElementById('convert-result'),
+  convertRows:          document.getElementById('convert-rows'),
+  convertSummary:       document.getElementById('convert-summary'),
+  convertCancelBtn:     document.getElementById('convert-cancel-btn'),
+  convertConfirmBtn:    document.getElementById('convert-confirm-btn'),
+  settingsBtn:          document.getElementById('settings-btn'),
+  settingsModal:        document.getElementById('settings-modal'),
+  settingsMaxInput:     document.getElementById('settings-max-mib'),
+  settingsMaxHint:      document.getElementById('settings-max-hint'),
+  settingsTimeInput:    document.getElementById('settings-timeout-min'),
+  settingsError:        document.getElementById('settings-error'),
+  settingsCancelBtn:    document.getElementById('settings-cancel-btn'),
+  settingsConfirmBtn:   document.getElementById('settings-confirm-btn'),
+};
+
+// type 필터 버튼들은 NodeList — toolbar 추출 시 router.js로 이동 예정.
+export const typeButtons = $.browseToolbar.querySelectorAll('.type-btn');
