+++ web/index.html
@@ -171,6 +171,6 @@
     </div>
   </div>
 
-  <script src="/app.js?v=28"></script>
+  <script type="module" src="/main.js?v=29"></script>
 </body>
 </html>
